+++ src/BaseSpecification.php
@@ -14,120 +14,15 @@ declare(strict_types=1);
 
 namespace Happyr\DoctrineSpecification;
 
-use Doctrine\ORM\QueryBuilder;
-use Happyr\DoctrineSpecification\Filter\Filter;
-use Happyr\DoctrineSpecification\Filter\Satisfiable;
-use Happyr\DoctrineSpecification\Query\QueryModifier;
-use Happyr\DoctrineSpecification\Specification\Specification;
+use Happyr\DoctrineSpecification\Specification\BaseSpecification as BaseBaseSpecification;
+
+@trigger_error('The '.__NAMESPACE__.'\BaseSpecification class is deprecated since version 1.1 and will be removed in 2.0, use \Happyr\DoctrineSpecification\Specification\BaseSpecification instead.', E_USER_DEPRECATED);
 
 /**
  * Extend this abstract class if you want to build a new spec with your domain logic.
+ *
+ * @deprecated This class is deprecated since version 1.1 and will be removed in 2.0, use \Happyr\DoctrineSpecification\Specification\BaseSpecification instead.
  */
-abstract class BaseSpecification implements Specification
+abstract class BaseSpecification extends BaseBaseSpecification
 {
-    /**
-     * @var string|null
-     */
-    private $context;
-
-    /**
-     * @param string|null $context
-     */
-    public function __construct($context = null)
-    {
-        $this->context = $context;
-    }
-
-    /**
-     * @param QueryBuilder $qb
-     * @param string       $context
-     *
-     * @return string
-     */
-    public function getFilter(QueryBuilder $qb, string $context): string
-    {
-        $spec = $this->getSpec();
-
-        if ($spec instanceof Filter) {
-            return $spec->getFilter($qb, $this->getContext($context));
-        }
-
-        return '';
-    }
-
-    /**
-     * @param QueryBuilder $qb
-     * @param string       $context
-     */
-    public function modify(QueryBuilder $qb, string $context): void
-    {
-        $spec = $this->getSpec();
-
-        if ($spec instanceof QueryModifier) {
-            $spec->modify($qb, $this->getContext($context));
-        }
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function filterCollection(iterable $collection): iterable
-    {
-        $spec = $this->getSpec();
-
-        if ($spec instanceof Satisfiable) {
-            return $spec->filterCollection($collection);
-        }
-
-        return $collection;
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function isSatisfiedBy($candidate): bool
-    {
-        $spec = $this->getSpec();
-
-        if ($spec instanceof Satisfiable) {
-            return $spec->isSatisfiedBy($candidate);
-        }
-
-        return true;
-    }
-
-    /**
-     * Return all the specifications.
-     *
-     * @return Filter|QueryModifier
-     */
-    abstract protected function getSpec();
-
-    /**
-     * @param string $context
-     *
-     * @return string
-     */
-    private function getContext(string $context): string
-    {
-        if (null !== $this->context) {
-            return sprintf('%s.%s', $context, $this->context);
-        }
-
-        return $context;
-    }
-
-    /**
-     * @param string $context
-     *
-     * @return string
-     */
-    protected function getNestedContext(string $context): string
-    {
-        if (null !== $this->context) {
-            return sprintf('%s.%s', $this->context, $context);
-        }
-
-        return $context;
-    }
 }
